+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Rename custom LSP methods names. https://github.com/rescript-lang/rescript-vscode/pull/611
 - Better performance for Inlay Hints and Codelens.
 - Accept both `@ns.doc` and the new `@res.doc` for the internal representation of doc comments. And both `@ns.optional` and `@res.optional` for the optional fields. https://github.com/rescript-lang/rescript-vscode/pull/642
+- Migrate `prepareRename` to analysis
 
 #### :bug: Bug Fix